Refactor: merge gis and rest since they are all generic REST calls.
