+++ art17/aggregation/agregator/primary.py
@@ -1,3 +1,6 @@
+"""
+    TODO: check usage and remove. Almost sure this isn't used any more.
+"""
 import flask
 import logging
 from art17 import models
